added change password option in user dashboard
