+++ src/components/Highlights/Highlights.scss
@@ -58,3 +58,10 @@
 .errorBox{
   color:red;
 }
+@media only screen and (max-width: 600px) {
+  .highlights .item{
+    height:50px;
+
+  }
+}
+
