+++ src/config.ts
@@ -1,4 +1,5 @@
-export const REACT_BACKEND_URL=process.env.REACT_APP_BACKEND_URL || 'http://localhost:3000';
+//export const REACT_BACKEND_URL=process.env.REACT_APP_BACKEND_URL || 'http://localhost:3000';
+export const REACT_BACKEND_URL='http://localhost:3000';
 export const REACT_STRIPE_PUBLIC_KEY=process.env.REACT_APP_STRIPE_PUBLIC_KEY;
 export const REACT_BUCKETNAME=process.env.REACT_APP_BUCKET_NAME || '';
 export const REACT_DIRNAME=process.env.REACT_APP_DIRNAME || '';
