+++ src/pages/Home/JobDetails.scss
@@ -235,6 +235,7 @@
 .companyLogoImage {
   background-position: center;
   border-radius: 20px;
+  height: 120px;
 }
 
 .jobTitle .closingDate{
