+++ src/pages/Home/JobFeed.scss
@@ -130,6 +130,6 @@
 
 @media only screen and (max-width: 1050px) {
     .jobFeed{
-        min-height: 350px;
+        min-height: 400px;
     }
 }
