+++ src/pages/SearchPage/SeachUtils/FullJobView.scss
@@ -26,6 +26,5 @@
 .searchedJob .job{
   padding:10px;
   border-radius: 6px;
-  width: max-content;
   margin: auto;
 }
