+++ src/pages/UserPage/User.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import './User.scss';
-import { HistoryJobs, ResetPassword } from './Tabs';
+import { HistoryJobs, ChangePassword } from './Tabs';
 
 const User=()=>{
   const [tab, setTab] = useState('reset');
@@ -11,19 +11,19 @@ const User=()=>{
           <h1>User Options</h1>
           <ul>
             <li 
-              className={tab==="history"?'active':''}
+              className={tab==='history'?'active':''}
               onClick={()=>setTab('history')}
-              >History
+            >History
             </li>
             <li
-              className={tab==="reset"?'active':''}
+              className={tab==='reset'?'active':''}
               onClick={()=>setTab('reset')}
-            >Reset Password
+            >Passwords
             </li>
           </ul>
         </div>
-        {tab=="history" && <HistoryJobs />}
-        {tab=="reset" && <ResetPassword />}
+        {tab=='history' && <HistoryJobs />}
+        {tab=='reset' && <ChangePassword />}
       </div>
     </div>
   );
